truck type add, delete and update  feature

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,8 +15,8 @@ class CreateUsersTable extends Migration
     {
         Schema::create('users', function (Blueprint $table) {
             $table->id();
-            $table->integer('role_id')->index()->unsigned()->nullable();
-            $table->integer('is_active')->default(0);
+            $table->bigInteger('role_id')->index()->unsigned()->nullable();
+            $table->bigInteger('is_active')->default(0);
             $table->string('name');
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
